Added inplace encryption/decryption for boxes
This is a low-level encryption/decryption API that doesn't do any memory
allocations, usable in low-level code. All functions that are part of
this api have names that end in `_inplace`.

All of the operations are carried out in place, however sodium has
input requirements (zero-padding) that needs to be fulfilled for these
functions to work. For that reason all of the low-level functions return
Option<&[u8]> to signal if the padding is valid. The functions modify
their input buffer and also return a slice pointing to the
plaintext/ciphertext portion of the buffer (i.e. not the padding).

Also rewrote the higher-level functions to use the low-level functions,
this simplified the higher-level functions a lot.